pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "strava-weekly-export"
version = "0.1.0"
description = "CLI tool to export Strava weekly activities to JSON"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.24",
  "python-dateutil>=2.8",
]

[project.scripts]
strava-weekly-export = "strava_export.cli:main"

[tool.setuptools.packages.find]
where = ["."]
